Add Ui class to handle user interface outputs
