+++ src/main/java/TaskFactory.java
@@ -38,7 +38,7 @@ public class TaskFactory {
         try {
             return Optional.of(constructTask(taskData));
         } catch (IllegalArgumentException ex) {
-            Duke.messagePrint("(>.<') did not understand this task - dropping it",
+            Ui.messagePrint("(>.<') did not understand this task - dropping it",
                     String.join(", ", taskData));
             return Optional.empty();
         }
